+++ shrdr/src/core/parallel_bk.h
@@ -36,7 +36,7 @@ namespace shrdr {
 
 using BlockIdx = uint16_t; // We assume 65536 is enough blocks
 using Time = uint32_t;
-using Dist = uint16_t;
+using Dist = uint32_t;
 
 enum NodeArcSorting : int32_t {
     LIFO = 0,
